modification pour machine 64

+++ audela/src/audela/libaudela/src/cpixels.h
@@ -76,8 +76,12 @@ public:
    virtual int  GetWidth(void)=0;
    virtual void GetPix(int *plane, TYPE_PIXELS *val1,TYPE_PIXELS *val2,TYPE_PIXELS *val3,int x, int y)=0;
    virtual void GetPixelsPointer(TYPE_PIXELS **pixels)=0;
-   virtual void GetPixels(int x1, int y1, int x2, int y2, TPixelFormat pixelFormat, TColorPlane plane,int pixels)=0;
-   virtual void GetPixelsReverse(int x1, int y1, int x2, int y2, TPixelFormat pixelFormat, TColorPlane plane, int pixels)=0;
+   //Yassine
+   //virtual void GetPixels(int x1, int y1, int x2, int y2, TPixelFormat pixelFormat, TColorPlane plane,int pixels)=0;
+   virtual void GetPixels(int x1, int y1, int x2, int y2, TPixelFormat pixelFormat, TColorPlane plane,long pixels)=0;
+   //Yassine
+   //virtual void GetPixelsReverse(int x1, int y1, int x2, int y2, TPixelFormat pixelFormat, TColorPlane plane, int pixels)=0;
+   virtual void GetPixelsReverse(int x1, int y1, int x2, int y2, TPixelFormat pixelFormat, TColorPlane plane, long pixels)=0;
    virtual void GetPixelsVisu( int x1,int y1,int x2, int y2,
                   int mirrorX, int mirrorY,
                    float *cuts,
